requirements.txt
fastapi
uvicorn
httpx
google-generativeai
python-dotenv
pydantic